requirement.txt
python-dotenv==1.0.0
pinecone-client==2.2.1
openai==0.27.8
uvicorn
langchain==0.0.230
uvicorn==0.22.0
lark
fastapi